new email checking impl
