+++ api/discord_email.go
@@ -9,6 +9,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"strings"
 )
 
 type discordResponse struct {
@@ -22,16 +23,21 @@ type discordResponse struct {
 	} `json:"errors"`
 }
 
-func Discord(mailService MailService, email string, config ApiConfig) (error, bool) {
+func DiscordMail(mailService MailService, email string, config ApiConfig) (EmailService, error) {
+	emailService := EmailService{
+		Name: mailService.Name,
+		Icon: mailService.Icon,
+		Link: strings.ReplaceAll(mailService.Url, "{{ email }}", email),
+	}
 	if config.Testing {
 		if email == "has_discord_account@gmail.com" || email == "discord@gmail.com" || email == "all@gmail.com" {
 			log.Println("has_discord_account testing case true")
-			return nil, true
+			return emailService, nil
 		} else if email == "discord_error@gmail.com" {
-			return errors.New("error"), false
+			return EmailService{}, errors.New("error")
 		}
 		log.Println("has_discord_account testing case false")
-		return nil, false
+		return EmailService{}, nil
 	}
 
 	log.Println("Checking Discord email")
@@ -43,7 +49,7 @@ func Discord(mailService MailService, email string, config ApiConfig) (error, bo
 	r, err := http.NewRequest("POST", endpoint, bytes.NewBuffer(jsonStr)) // URL-encoded payload
 	if err != nil {
 		log.Println(err)
-		return err, false
+		return EmailService{}, err
 	}
 	r.Header.Add("Content-Type", "application/json")
 	r.Header.Add("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.149 Safari/537.36")
@@ -52,7 +58,7 @@ func Discord(mailService MailService, email string, config ApiConfig) (error, bo
 	res, err := client.Do(r)
 	if err != nil {
 		log.Println(err)
-		return err, false
+		return EmailService{}, err
 	}
 	defer res.Body.Close()
 	if res.StatusCode == 400 {
@@ -61,13 +67,13 @@ func Discord(mailService MailService, email string, config ApiConfig) (error, bo
 		json.Unmarshal(body, &response)
 		if len(response.Errors.Email.Errors) > 0 {
 			if response.Errors.Email.Errors[0].Code == "EMAIL_ALREADY_REGISTERED" {
-				return nil, true
+				return emailService, nil
 			}
 		}
 	} else if res.StatusCode == 429 {
 		//("Too many requests to Discord!")
 		log.Println("to many requests")
-		return errors.New("to many requests"), false
+		return EmailService{}, errors.New("to many requests")
 	}
-	return nil, false
+	return EmailService{}, nil
 }
