+++ api/twitter_email.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"net/http"
 	"net/url"
+	"strings"
 )
 
 type twitterResponse struct {
@@ -14,14 +15,20 @@ type twitterResponse struct {
 	Taken bool   `json:"taken"`
 }
 
-func Twitter(mailService MailService, email string, config ApiConfig) (error, bool) {
+func TwitterMail(mailService MailService, email string, config ApiConfig) (EmailService, error) {
+	emailService := EmailService{
+		Name: mailService.Name,
+		Icon: mailService.Icon,
+		Link: strings.ReplaceAll(mailService.Url, "{{ email }}", email),
+	}
 	if config.Testing {
 		if email == "has_twitter_account@gmail.com" || email == "twitter@gmail.com" || email == "all@gmail.com" {
 			log.Println("has_twitter_account testing case true")
-			return nil, true
+			return emailService, nil
 		}
 		log.Println("has_twitter_account testing case false")
-		return nil, false
+
+		return EmailService{}, nil
 	}
 	var endpoint = "https://api.twitter.com/i/users/email_available.json"
 
@@ -31,23 +38,27 @@ func Twitter(mailService MailService, email string, config ApiConfig) (error, bo
 	r, err := http.Get(endpoint + "?" + data.Encode())
 	if err != nil {
 		log.Println(err)
-		return err, false
+
+		return EmailService{}, err
 	}
 	r.Header.Add("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.149 Safari/537.36")
 	if err != nil {
 		log.Println(err)
-		return err, false
+
+		return EmailService{}, err
 	}
 	if r.StatusCode == 200 {
 		body, _ := ioutil.ReadAll(r.Body)
 		var response twitterResponse
 		json.Unmarshal(body, &response)
 		if response.Taken {
-			return nil, true
+			return emailService, nil
 		} else {
-			return nil, false
+
+			return EmailService{}, nil
 		}
 	} else {
-		return nil, false
+
+		return EmailService{}, nil
 	}
 }
